requirements/base.txt
pandas==2.2.2
numpy==2.1.2
matplotlib==3.10.1
seaborn==0.13.2
adjusttext==1.3.0
scikit-learn==1.4.2
flax==0.10.4
jax==0.5.1
jaxlib==0.5.1
triton<3
ml-dtypes==0.5.1
tensorstore==0.1.72
orbax-checkpoint==0.11.9
optax==0.2.4
click<8.2
typer==0.7.0
ipinfo==5.1.1
google-cloud-storage==3.0.0
types-tqdm==4.67.0.20250516
types-tensorflow==2.18.0.20250516
pandas-stubs==2.2.3.250527
types-seaborn==0.13.2.20250516
types-networkx==3.5.0.20250610
scipy-stubs==1.15.3.0
